requirements_old.txt
absl-py==0.9.0
asgiref==3.2.7
astunparse==1.6.3
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
confusable-homoglyphs==3.2.0
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.7.0rc1
django-admin-tools==0.9.0
django-crispy-forms==1.9.0
django-registration==3.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
Markdown==3.2.2
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
protobuf==3.11.3
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
textblob==0.15.3
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
